+++ examples/data-sources/nxos_ipv6_interface/data-source.tf
@@ -1,4 +1,0 @@
-data "nxos_ipv6_interface" "example" {
-  vrf          = "default"
-  interface_id = "eth1/10"
-}