+++ examples/data-sources/nxos_ipv6_interface_address/data-source.tf
@@ -1,5 +1,0 @@
-data "nxos_ipv6_interface_address" "example" {
-  vrf          = "default"
-  interface_id = "eth1/10"
-  address      = "2001:db8:3333:4444:5555:6666:7777:8888"
-}