+++ examples/resources/nxos_ipv6_interface/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_ipv6_interface.example "<vrf>,<interface_id>"