+++ examples/resources/nxos_ipv6_interface_address/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_ipv6_interface_address.example
-  id = "<vrf>,<interface_id>,<address>"
-}