+++ examples/resources/nxos_ipv6_interface_address/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_ipv6_interface_address.example "<vrf>,<interface_id>,<address>"