+++ examples/resources/nxos_ipv6_static_route/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_ipv6_static_route.example
-  id = "<vrf_name>,<prefix>"
-}