+++ examples/resources/nxos_ipv6_static_route/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_ipv6_static_route.example "<vrf_name>,<prefix>"