+++ docs/guides/changelog.md
@@ -35,6 +35,7 @@ description: |-
 - BREAKING CHANGE: Remove `nxos_ospfv3_instance` resource and data source, `nxos_ospfv3_vrf` resource and data source, `nxos_ospfv3_area` resource and data source, `nxos_ospfv3_vrf_address_family` resource and data source, and `nxos_ospfv3_interface` resource and data source, OSPFv3 configuration is now managed as a single `nxos_ospfv3` resource and data source
 - BREAKING CHANGE: Remove `nxos_ipv4_prefix_list_rule` resource and data source and `nxos_ipv4_prefix_list_rule_entry` resource and data source, IPv4 Prefix List configuration is now managed as a single `nxos_ipv4_prefix_list` resource and data source
 - BREAKING CHANGE: Remove `nxos_ipv4_static_route` resource and data source, `nxos_ipv4_interface` resource and data source, and `nxos_ipv4_interface_address` resource and data source, IPv4 configuration is now managed as a single `nxos_ipv4_vrf` resource and data source
+- BREAKING CHANGE: Remove `nxos_ipv6_static_route` resource and data source, `nxos_ipv6_interface` resource and data source, and `nxos_ipv6_interface_address` resource and data source, IPv6 configuration is now managed as a single `nxos_ipv6_vrf` resource and data source
 
 ## 0.7.0